20161226 - implement ACL (roles and permissions, starting with management of attachments)

+++ app/Http/Controllers/AttachmentsController.php
@@ -8,7 +8,7 @@ use Illuminate\Support\Facades\File;
 use Response;
 use Image;
 use Illuminate\Support\Facades\Redirect;
-
+use Gate;
 
 class AttachmentsController extends Controller
 {
@@ -27,32 +27,39 @@ class AttachmentsController extends Controller
         return $sanitized;
     }
     public function show_attachment($entity, $entity_id, $type='attachment', $file_name=NULL)
-    {
+    {   
         switch ($type) {
             case 'attachment' :
+                $this->authorize('show-attachment');
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/attachments/'.$file_name;
                 break;
             case 'avatar' :
+                $this->authorize('show-avatar');
                 $file_name = 'avatar.png';
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/'.$file_name;
                 break;
             case 'schedule' :
+                $this->authorize('show-attachment');
                 $file_name = 'schedule.pdf';
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/'.$file_name;
                 break;
             case 'contract' :
+                $this->authorize('show-attachment');
                 $file_name = 'contract.pdf';
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/'.$file_name;
                 break;
             case 'evaluations' :
+                $this->authorize('show-attachment');
                 $file_name = 'evaluations.pdf';
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/'.$file_name;
                 break;
             case 'group_photo' :
+                $this->authorize('show-attachment');
                 $file_name = 'group_photo.jpg';
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/'.$file_name;
                 break;
             default :
+                $this->authorize('show-attachment');
                 $path = storage_path() . '/app/'.$entity.'/'.$entity_id. '/'.$file_name;
                 break;
         }
@@ -66,63 +73,75 @@ class AttachmentsController extends Controller
         $response->header("Content-Type", $type);
 
         return $response;
+        
+         
     }
 
     public function store_attachment($file, $entity='event',$entity_id=0, $type=NULL,$description=NULL) {
-            $file_name = $this->sanitize_filename($file->getClientOriginalName());
-            $attachment = new \montserrat\Attachment;
-            $attachment->mime_type = $file->getClientMimeType();
-            $attachment->description = $description;
-            $attachment->upload_date = \Carbon\Carbon::now();
-            $attachment->entity = $entity;
-            $attachment->entity_id = $entity_id;
-            switch ($type) {
-                case 'contract': 
-                    $attachment->file_type_id = FILE_TYPE_EVENT_CONTRACT;
-                    $attachment->uri = 'contract.pdf';
-                    break;
-                case 'schedule' :
-                    $attachment->file_type_id = FILE_TYPE_EVENT_SCHEDULE;
-                    $attachment->uri = 'schedule.pdf';
-                    break;
-                case 'evaluation' :
-                    $attachment->file_type_id = FILE_TYPE_EVENT_EVALUATION;
-                    $attachment->uri = 'evaluations.pdf';
-                    break;
-                case 'group_photo' :
-                    $attachment->file_type_id = FILE_TYPE_EVENT_GROUP_PHOTO;
-                    $attachment->uri = 'group_photo.png';
-                    break;
-                case 'attachment' :
-                    $attachment->file_type_id = FILE_TYPE_CONTACT_ATTACHMENT;
-                    $attachment->uri = $file_name;
-                    break;
-                case 'avatar' :
-                    $avatar = Image::make($file->getRealPath())->fit(150, 150)->orientate();
-                    $attachment->file_type_id = FILE_TYPE_CONTACT_AVATAR;
-                    $attachment->uri = 'avatar.png';
-                    break;
-                default : 
-                    $attachment->file_type_id = 0;
-                    $attachment->uri = $file_name;
-                    break;
-            }
-            $attachment->save();
-            //write file to filesystem
-            if ($type=='avatar') {
-                Storage::disk('local')->put($entity.'/'.$entity_id.'/'.'avatar.png',$avatar->stream('png'));
-
-            } else {
-                Storage::disk('local')->put($entity.'/'.$entity_id.'/'.$attachment->uri,File::get($file));
-            }
+        $this->authorize('create-attachment');
+        $file_name = $this->sanitize_filename($file->getClientOriginalName());
+        $attachment = new \montserrat\Attachment;
+        $attachment->mime_type = $file->getClientMimeType();
+        $attachment->description = $description;
+        $attachment->upload_date = \Carbon\Carbon::now();
+        $attachment->entity = $entity;
+        $attachment->entity_id = $entity_id;
+        switch ($type) {
+            case 'contract': 
+                $this->authorize('create-event-contract');
+                $attachment->file_type_id = FILE_TYPE_EVENT_CONTRACT;
+                $attachment->uri = 'contract.pdf';
+                break;
+            case 'schedule' :
+                $this->authorize('create-event-schedule');
+                $attachment->file_type_id = FILE_TYPE_EVENT_SCHEDULE;
+                $attachment->uri = 'schedule.pdf';
+                break;
+            case 'evaluation' :
+                $this->authorize('create-event-evaluation');
+                $attachment->file_type_id = FILE_TYPE_EVENT_EVALUATION;
+                $attachment->uri = 'evaluations.pdf';
+                break;
+            case 'group_photo' :
+                $this->authorize('create-event-group-photo');
+                $attachment->file_type_id = FILE_TYPE_EVENT_GROUP_PHOTO;
+                $attachment->uri = 'group_photo.png';
+                break;
+            case 'attachment' :
+                $this->authorize('create-attachment');
+                $attachment->file_type_id = FILE_TYPE_CONTACT_ATTACHMENT;
+                $attachment->uri = $file_name;
+                break;
+            case 'avatar' :
+                $this->authorize('create-avatar');
+                $avatar = Image::make($file->getRealPath())->fit(150, 150)->orientate();
+                $attachment->file_type_id = FILE_TYPE_CONTACT_AVATAR;
+                $attachment->uri = 'avatar.png';
+                break;
+            default : 
+                $this->authorize('create-attachment');
+                $attachment->file_type_id = 0;
+                $attachment->uri = $file_name;
+                break;
         }
+        $attachment->save();
+        //write file to filesystem
+        if ($type=='avatar') {
+            Storage::disk('local')->put($entity.'/'.$entity_id.'/'.'avatar.png',$avatar->stream('png'));
+
+        } else {
+            Storage::disk('local')->put($entity.'/'.$entity_id.'/'.$attachment->uri,File::get($file));
+        }
+    }
 
     public function update_attachment ($file, $entity='event',$entity_id=0, $type=NULL,$description=NULL) {
-
+        
+               
         $path = $entity.'/'.$entity_id.'/';
 
         switch ($type) {
             case 'avatar' : 
+                $this->authorize('create-avatar'); //if you can create it you can update it 
                 $file_type_id = FILE_TYPE_CONTACT_AVATAR;
                 $file_name = 'avatar.png';
                 $updated_file_name= 'avatar-updated-'.time().'.png';
@@ -138,6 +157,7 @@ class AttachmentsController extends Controller
 
                 break;
             case 'attachment' :
+                $this->authorize('create-attachment');
                 $path = $entity.'/'.$entity_id.'/attachments/';
                 $file_type_id = FILE_TYPE_CONTACT_ATTACHMENT;
                 $file_name = $this->sanitize_filename($file->getClientOriginalName());
@@ -152,6 +172,7 @@ class AttachmentsController extends Controller
                 }
                 break;
             case 'schedule' :
+                $this->authorize('create-event-schedule');
                 $path = $entity.'/'.$entity_id.'/';
                 $file_type_id = FILE_TYPE_EVENT_SCHEDULE;
                 $file_name = 'schedule.pdf';
@@ -165,6 +186,7 @@ class AttachmentsController extends Controller
                 }
                 break;
             case 'contract' :
+                $this->authorize('create-event-contract');
                 $path = $entity.'/'.$entity_id.'/';
                 $file_type_id = FILE_TYPE_EVENT_SCHEDULE;
                 $file_name = 'contract.pdf';
@@ -178,6 +200,7 @@ class AttachmentsController extends Controller
                 }
                 break;
             case 'evaluations' :
+                $this->authorize('create-event-evaluation');
                 $path = $entity.'/'.$entity_id.'/';
                 $file_type_id = FILE_TYPE_EVENT_EVALUATION;
                 $file_name = 'evaluations.pdf';
@@ -191,6 +214,7 @@ class AttachmentsController extends Controller
                 }
                 break;
             case 'group_photo' :
+                $this->authorize('create-event-group-photo');
                 $path = $entity.'/'.$entity_id.'/';
                 $file_type_id = FILE_TYPE_EVENT_GROUP_PHOTO;
                 $file_name = 'group_photo.jpg';
@@ -205,6 +229,7 @@ class AttachmentsController extends Controller
                 break;
                 
             default :
+                $this->authorize('create-attachment');
                 break;
         }
         $attachment = \montserrat\Attachment::firstOrNew(['entity'=>$entity,'entity_id'=>$entity_id,'file_type_id'=>$file_type_id,'uri'=>$file_name]);
@@ -216,6 +241,8 @@ class AttachmentsController extends Controller
         $attachment->save();
     }
     public function delete_attachment($file_name, $entity='event',$entity_id=0, $type=NULL) {
+        $this->authorize('delete-attachment');
+               
         $path = $entity.'/'.$entity_id.'/';
         switch ($type) {
             case 'group_photo' :
@@ -268,63 +295,71 @@ class AttachmentsController extends Controller
     }
 
     public function show_contact_attachment($user_id, $file_name)
-    {
+    {   $this->authorize('show-attachment');
         return $this->show_attachment('contact',$user_id,'attachment',$file_name);
     
     }
     public function delete_contact_attachment($user_id, $attachment)
-    {
+    {   $this->authorize('delete-attachment');
         $this->delete_attachment($attachment,'contact',$user_id,'attachment');
         // TODO: get contact type and redirect to person, parish, organization, vendor as appropriate
         return Redirect::action('PersonsController@show',$user_id);
         
     }    
     
-    public function get_avatar($user_id)
-    {
+    public function get_avatar($user_id) {   
+        $this->authorize('show-avatar');
         return $this->show_attachment('contact',$user_id,'avatar','avatar.png');
     }
     
     public function delete_avatar($user_id)
-    {
+    {   $this->authorize('delete-attachment');
         $this->delete_attachment('avatar.png','contact',$user_id,'avatar');
         return Redirect::action('PersonsController@show',$user_id);
         
     }
 
     public function get_event_contract($event_id) {
+        $this->authorize('show-event-attachment');
         return $this->show_attachment('event',$event_id,'contract',NULL);
     }    
 
     public function get_event_schedule($event_id) {
+        $this->authorize('show-event-attachment');
         return $this->show_attachment('event',$event_id,'schedule',NULL);
     }    
 
     
     public function get_event_evaluations($event_id) {
+        $this->authorize('show-event-evaluation');
         return $this->show_attachment('event',$event_id,'evaluations',NULL);
     }  
     
     public function delete_event_evaluations($event_id) {
+        $this->authorize('delete-attachment');
         $this->delete_attachment('evaluations.pdf','event',$event_id,'evaluations');
         return Redirect::action('RetreatsController@show',$event_id);
     } 
     public function delete_event_schedule($event_id) {
+        $this->authorize('delete-attachment');
         $this->delete_attachment('schedule.pdf','event',$event_id,'schedule');
         return Redirect::action('RetreatsController@show',$event_id);
     }
 
     public function delete_event_contract($event_id) {
+        $this->authorize('delete-attachment');
         $this->delete_attachment('contract.pdf','event',$event_id,'contract');
         return Redirect::action('RetreatsController@show',$event_id);
     }
     
     public function delete_event_group_photo($event_id) {
+        $this->authorize('delete-attachment');
            $this->delete_attachment('group_photo.jpg','event',$event_id,'group_photo');
            return Redirect::action('RetreatsController@show',$event_id);
     }
 
     public function get_event_group_photo($event_id) {
+        $this->authorize('show-event-photo');
         return $this->show_attachment('event',$event_id,'group_photo',NULL);
     
     }
